+++ packages/toggle/src/Toggle.js
@@ -52,8 +52,10 @@ class Toggle extends Component<Props> {
 		uncheckedButtonClass: 'a-button--danger'
 	}
 
-  handleClick = e => {
-    this.setState({ checked: e.target.value });
+  handleChange = e => {
+    if (this.props.onChange) {
+      this.props.onChange(e);
+    }
     if (this.props.onClick) {
       this.props.onClick(e);
     }
@@ -103,11 +105,11 @@ class Toggle extends Component<Props> {
           className="a-toggle__checkbox"
           id={id}
           name={name}
-          aria-checked={this.state.checked}
-          defaultChecked={this.state.checked}
+          aria-checked={this.props.checked}
+          defaultChecked={this.props.checked}
           type="checkbox"
           role="switch"
-          onChange={this.handleClick}
+          onChange={this.handleChange}
         />
         <div className="a-toggle__labels">
           <label htmlFor={id} className={checkedClasses}>